Add new meson options to suppress compilation of daemon and examples
This makes it quicker and easier to build multilib + multiarch variants
of the library without having to compile the larget host-arch specific
daemon. This is required for using the automatic LD_PRELOAD module with
older 32-bit games on a 64-bit host, to ensure they all have access to
the library.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>